rename react-hig to hig-react and remove private repo
