+++ src/implementations/react/package.json
@@ -1,9 +1,9 @@
 {
-  "name": "@hig/react-hig",
+  "name": "hig-react",
   "version": "1.0.3",
   "description": "hig components in react",
-  "main": "lib/react-hig.js",
-  "css": "lib/react-hig.css",
+  "main": "lib/hig-react.js",
+  "css": "lib/hig-react.css",
   "author": "Autodesk Inc.",
   "homepage": ".",
   "license": "Apache-2.0",
@@ -45,7 +45,7 @@
     "build": "npm run build-vanilla && NODE_ENV=production webpack -p",
     "build-playground": "react-scripts build",
     "build-storybook": "build-storybook -s public -o build/storybook",
-    "build-vanilla": "cd ../vanilla && npm run build && cd ../react && mkdir -p lib && cp ../vanilla/dist/hig.css lib/react-hig.css",
+    "build-vanilla": "cd ../vanilla && npm run build && cd ../react && mkdir -p lib && cp ../vanilla/dist/hig.css lib/hig-react.css",
     "playground": "react-scripts start",
     "storybook": "start-storybook -p 9009 -s public",
     "start": "concurrently 'npm run start-vanilla' 'webpack --watch' --kill-others",
