+++ src/implementations/react/src/playground/sections/RangeSection.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PlaygroundSection from '../PlaygroundSection';
-import { Range } from '../../react-hig';
+import { Range } from '../../hig-react';
 
 function RangeSection() {
   return (
