+++ src/implementations/react/src/playground/sections/TypographySection.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PlaygroundSection from '../PlaygroundSection';
-import { H1, H2, H3, Sub1, Sub2, Body, Bold, Caption, Disabled } from '../../react-hig'
+import { H1, H2, H3, Sub1, Sub2, Body, Bold, Caption, Disabled } from '../../hig-react'
 import createSlotComponent from '../../adapters/createSlotComponent';
 
 const Slot = createSlotComponent();
