+++ src/implementations/react/src/react-hig.test.js
@@ -1,9 +1,9 @@
-import * as ReactHIG from './react-hig';
+import * as ReactHIG from './hig-react';
 
 import Button from './adapters/ButtonAdapter';
 import GlobalNav from './elements/components/GlobalNav/GlobalNav';
 
-describe('react-hig', () => {
+describe('hig-react', () => {
   it('exports Button', () => {
     expect(Button).toEqual(ReactHIG.Button);
   });
