+++ src/implementations/vanilla/package.json
@@ -24,9 +24,9 @@
   "author": "Autodesk Inc.",
   "license": "Apache 2.0",
   "devDependencies": {
-    "babel-core": "^6.23.1",
+    "babel-core": "^6.26.0",
+    "babel-loader": "^7.1.1",
     "babel-eslint": "^7.2.3",
-    "babel-loader": "^6.3.0",
     "babel-plugin-module-resolver": "^2.7.1",
     "babel-plugin-syntax-dynamic-import": "^6.18.0",
     "babel-preset-es2015": "^6.22.0",
@@ -53,6 +53,6 @@
     "sass-loader": "^6.0.0",
     "style-loader": "^0.13.1",
     "svgo": "^0.7.2",
-    "webpack": "^2.2.1"
+    "webpack": "^3.5.5"
   }
 }
